If SLA tracking is not active do not create meta key
Previously the `_kbs_ticket_sla_target_respond` and
`_kbs_ticket_sla_target_resolve` meta keys were always created, even
with SLA tracking disabled. This resulted in a target date of 1970-01-01
00:00:00 for both keys.
With this change, if SLA tracking is disabled, no SLA meta keys will be
created. Also allows for simpler checking on a per ticket basis as to
whether or not the SLA for the ticket should be tracked.