Ensure agents have access to all articles regardless of restrictions
Within core KBS, this function does not have much value as a restricted
article is only restricted to users who are not logged in, however with
extensions these restrictions are more defined. For example, with the
EDD and WooCommerce extension for KBS, restrictions may be applied to
users who have purchased a specific product

With this function, we ensure that all agents have access to every
article regardless of the restrictions in place.