Remove `{agent}` tag if no agent assigned. Fixes #132
The `{agent}` template tag can be used within the **Ticket Notification Emails** setting. When ticket notifications are being generated, the tag is replaced with the email address of the primary agent of the ticket.

However, if auto assignment is not enabled and no agent is currently assigned to the ticket, the `{agent}` template tag cannot be replaced and it causes an email failure message.

To resolve this, we need to check the recipient string for the presence of `{agent}` and then flag if it is not possible to replace with an email address. When this flag is identified, we simply remove `{agent}` from the recipient string.